refactor: move metric endpoint to goals feature folder
