+++ api/src/features/goal/routes.py
@@ -6,6 +6,8 @@ from sqlalchemy.engine import Connection
 from src.core.auth import get_current_student_id
 from src.core.db import get_connection
 from src.core.logger import get_logger
+from src.features.analytics import service as analytics_service
+from src.features.analytics.schemas import StrategyMetric, StrategyMetricsResponse
 from src.features.study_session import service as study_session_service
 from src.shared.schemas import Status
 
@@ -80,3 +82,24 @@ def delete_goal(
 
     if not success:
         raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail='Goal not found')
+
+
+@router.get('/{goal_id}/strategy-metrics', response_model=StrategyMetricsResponse)
+def get_goal_strategy_metrics(
+    goal_id: str,
+    conn: Connection = Depends(get_connection),
+    student_id: int = Depends(get_current_student_id),
+):
+    strategy_metrics = analytics_service.get_strategy_metrics(conn, student_id, goal_id)
+
+    return StrategyMetricsResponse(
+        goal_id=goal_id,
+        strategy_metrics=[
+            StrategyMetric(
+                strategy=strategy,
+                adherence=data['adherence'],
+                sessions_count=data['sessions_count'],
+            )
+            for strategy, data in strategy_metrics.items()
+        ],
+    )
