+++ bruno/Ownit/environments/study_session_id.bru
@@ -1,3 +1,4 @@
 vars {
   study_session_id: 9ecbfb1f-e0ba-4bf2-b811-656a45bbc6c3
+  goal_id: 50afa447-748d-4512-9c73-33dda6be7b62
 }
